Original code P_2_1_2_04
(My poor attempt at explaining the code in English)

When the mouse is at (0,0), the tileCount and rectSize are regular and
thus the whole screen is covered in rectangles hitting each other along
the edge. The vertices of the rectangle shift/distort in relation to
the position of the mouse. When the mouse moves along the edge of the
Y-axis the vertices of the rectangle remain in line with the Y
co-ordinate and visa versa when the mouse moves along the X-axis. When
the mouse moves away from the top left corner and towards the bottom
right corner the rectangles expand out of the grid.

When the mouse is pressed the output randomises the rectangles in
relation to where the mouse is positioned.