Change grid position
The gridX and gridY commands ensured that the shapes would repeat along
the X and Y axis covering the entire window with an equal distance of 1
added in between each shape. By multiplying by 10 instead of gridY this
restricts the shapes to be output into a single line along the X axis
only. And the opposite would happen if 10 was multiplied instead of
gridX.