Change angle and radius
This code outputs a perfect regular circle from square lines.
A circle is created using the cos and sin values, changing the radius,
in this case multiplying by 2 and subtracting by 2, changes the size of
the circle.
Additionally the sin value has been replaced with tan to create a more
interesting shape.